requirements.txt
alpaca-py>=0.28.0
pandas>=2.0.0
numpy>=2.0.0
yfinance>=0.2.30
python-dotenv>=1.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=2.0.0
schedule>=1.2.0
PyQt5>=5.15.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
textblob>=0.18.0
requests>=2.31.0

